+++ src/app/admin-module/admin.module.ts
@@ -7,9 +7,11 @@ import { AdminRoutingModule } from './admin-routing.module';
 import { AutocompleteModule } from '../component-modules/autocomplete/autocomplete.module';
 import { BreadcrumbComponent } from './components/breadcrumb.component';
 import { DashboardComponent } from './components/dashboard.component';
-import { RecipeDetailCheckboxListComponent } from "./components/recipe-detail-checkbox-list.component";
+import { RecipeDetailCheckboxListComponent } from './components/recipe-detail-checkbox-list.component';
+import { RecipeIngredientDialogComponent } from './components/recipe-ingredient-dialog/recipe-ingredient-dialog.component';
 import { RecipeListComponent } from './components/recipe-list.component';
-import { RecipeDetailComponent } from './components/recipe-detail.component';
+import { RecipeDetailComponent } from './components/recipe-detail/recipe-detail.component';
+import { DataCollectionModule } from '../data-collection/data-collection.module';
 import { RecipeTitleDialogComponent } from './components/recipe-title-dialog.component';
 import { SourceDetailComponent } from './components/source-detail.component';
 import { SourceDialogComponent } from './components/source-dialog.component';
@@ -29,7 +31,8 @@ import { SharedMaterialModule } from '../shared/shared-material.module';
     SourcePageDialogComponent,
     BreadcrumbComponent,
     RecipeTitleDialogComponent,
-    RecipeDetailCheckboxListComponent
+    RecipeDetailCheckboxListComponent,
+    RecipeIngredientDialogComponent,
   ],
   imports: [
     ReactiveFormsModule,
@@ -37,6 +40,7 @@ import { SharedMaterialModule } from '../shared/shared-material.module';
     AdminRoutingModule,
     CommonModule,
     AutocompleteModule,
+    DataCollectionModule,
     SharedModule,
   ],
   providers: [DataCollectionService],