+++ src/app/admin-module/models/data-collection-detail.interface.ts
@@ -1,12 +1,12 @@
-import { Ingredient } from "../../search-module/models/ingredient.interface";
 import { Localized } from "../../shared/models/localized.type";
 
 export interface DataCollectionDetail {
   title: string;
   url: string;
   calories: string;
+  cookingTime: string | null;
   titleLocale: Localized | null;
   ingredients: string[];
-  requiredIngredients: Ingredient[];
-  optionalIngredients: Ingredient[];
+  requiredIngredients: number[];
+  optionalIngredients: number[];
 }