+++ src/app/admin-module/models/data-field-custom-action.type.ts
@@ -1,0 +1,4 @@
+export type DataFieldCustomAction = {
+  label: string;
+  onClick: () => void;
+};